+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.1.0 
 
+* Ignore finder preview files on the card
 * Removed dng conversion, Rafer stores raf and jpg now.
 * You can run rafer without a card and/or without available backups now.
 * Store artist and copyright.